+++ app/opulence/orm/ComparatorRegistry.php
@@ -1,0 +1,12 @@
+<?php
+/**
+ * Copyright (C) 2015 David Young
+ *
+ * Defines the registry for comparison functions
+ */
+namespace Opulence\ORM;
+
+class ComparatorRegistry
+{
+
+}