+++ app/opulence/orm/ids/IdGenerator.php
@@ -7,18 +7,17 @@
 namespace Opulence\ORM\Ids;
 
 use Opulence\Databases\IConnection;
-use Opulence\ORM\IEntity;
 
 abstract class IdGenerator
 {
     /**
      * Generates an Id for an entity
      *
-     * @param IEntity $entity The entity whose Id we're generating
+     * @param object $entity The entity whose Id we're generating
      * @param IConnection $connection The connection to use to get the Id
      * @return mixed The Id of the entity
      */
-    abstract public function generate(IEntity $entity, IConnection $connection);
+    abstract public function generate($entity, IConnection $connection);
 
     /**
      * Gets the value of the Id when it isn't set